# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: SourceCraft Repo Health API
  version: 1.0.0
  description: 'API сервиса SourceCraft Repo Health. Файл сгенерирован из кода (app.openapi()),
    живая версия - /api/docs. Ошибки: {"error": {"code", "message"}}. Анализ асинхронный:
    202 + job_id, статус - GET /api/v1/jobs/{job_id}. scope=public - публичный анализ
    (рейтинг), scope=personal - личный по PAT пользователя.'
paths:
...
  /api/v1/user/token:
    put:
      tags:
      - User
      summary: Set Token
      description: Проверяем токен запросом GET /user и сохраняем только в зашифрованном
        виде.
      operationId: set_token_api_v1_user_token_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - User
      summary: Delete Token
      operationId: delete_token_api_v1_user_token_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
  /api/v1/user/orgs:
    put:
      tags:
      - User
      summary: Set Orgs
      operationId: set_orgs_api_v1_user_orgs_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgsRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Me:
      properties:
        id:
          type: string
          title: Id
        login:
          anyOf:
          - type: string
          - type: 'null'
          title: Login
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        has_token:
          type: boolean
          title: Has Token
        sourcecraft_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Sourcecraft Username
        orgs:
          items:
            type: string
          type: array
          title: Orgs
      type: object
      required:
      - id
      - login
      - display_name
      - avatar_url
      - has_token
      - sourcecraft_username
      - orgs
      title: Me
...
    OrgsRequest:
      properties:
        orgs:
          items:
            type: string
          type: array
          title: Orgs
      type: object
      required:
      - orgs
      title: OrgsRequest
...
    TokenRequest:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
      - token
      title: TokenRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError